dataflow/1_BasicExample/requirements.txt
setuptools
wheel
apache_beam
apache-beam[gcp]
apache-beam[interactive]
avro
fastavro
google-cloud-storage
google-cloud-pubsub